reader/desktop/requirements.txt
git+https://github.com/nathanael6556/py-uhf-reader.git
paho-mqtt
protobuf
python-dotenv